Modelcreation

+++ src/main/java/de/hhu/propra2/material2/mops/models/Tag.java
@@ -1,16 +1,13 @@
 package de.hhu.propra2.material2.mops.models;
 
-import de.hhu.propra2.material2.mops.DTOs.DateiDTO;
 import lombok.Value;
-
-import java.util.List;
 
 @Value
 public class Tag {
     /**
      * Unique ID from database.
      */
-    private final int id;
+    private final long id;
     /**
      * The tags text that will be shown
      * and can be searched for.
